+++ src/api/Api.jsx
@@ -1,17 +1,5 @@
 import * as Mock from "./mocked-data.jsx";
 
-export function apiCallToGetTribes(areaId) {
-  return Mock.Tribes;
-}
-
-export function apiCallToGetTeams(tribeId) {
-  return Mock.Teams;
-}
-
-export function apiCallToGetEmployees(teamId) {
-  return Mock.Employees;
-}
-
 export function apiCallToGetEmployeeMetric(employeeId) {
   return Mock.Employees[0];
 }
@@ -56,6 +44,39 @@ export async function apiCallToGetAreas(companyId, companyName) {
 
   if (!response.ok) {
     throw new Error("Failed to get areas.");
+  }
+
+  return resData;
+}
+
+export async function apiCallToGetTribes(areaId, areaName) {
+  const response = await fetch('http://localhost:8080/tribe/tribes?id='+areaId+'&area='+areaName);
+  const resData = await response.json();
+
+  if (!response.ok) {
+    throw new Error("Failed to get tribes.");
+  }
+
+  return resData;
+}
+
+export async function apiCallToGetTeams(tribeId, tribeName) {
+  const response = await fetch('http://localhost:8080/team/teams?id='+tribeId+'&tribe='+tribeName);
+  const resData = await response.json();
+
+  if (!response.ok) {
+    throw new Error("Failed to get teams.");
+  }
+
+  return resData;
+}
+
+export async function apiCallToGetEmployees(teamId, teamName) {
+  const response = await fetch('http://localhost:8080/employee/allByTeam?id='+teamId+'&team='+teamName);
+  const resData = await response.json();
+
+  if (!response.ok) {
+    throw new Error("Failed to get employees.");
   }
 
   return resData;